+++ Board.h
@@ -8,18 +8,18 @@ class Board
 private:
     // if I allocated this on the heap, I could get the size right with resize
     std::vector<Cell> _board;
-    int _width = 0;
-    int _height = 0;
-    int _size = 0;
-    int _generation = 0;
+    uint16_t _width = 0;
+    uint16_t _height = 0;
+    uint16_t _size = 0;
+    uint16_t _generation = 0;
 
-    int _numDead = 0;
-    int _numLive = 0;
-    int _numBorn = 0;
-    int _numOld = 0;
-    int _numDying = 0;
-    int _OldAge = -1;
-    int _dirty = 0;
+    uint16_t _numDead = 0;
+    uint16_t _numLive = 0;
+    uint16_t _numBorn = 0;
+    uint16_t _numOld = 0;
+    uint16_t _numDying = 0;
+    uint16_t _OldAge = -1;
+    uint16_t _dirty = 0;
 
 public:
     explicit Board(std::nullptr_t) {};
@@ -27,9 +27,9 @@ public:
 
     ~Board() = default;
 
-    Board(int width, int height);
+    Board(uint16_t width, uint16_t height);
 
-    void SetOldAge(int age)
+    void SetOldAge(uint16_t age)
     {
         _OldAge = age;
     }
@@ -101,7 +101,7 @@ public:
 
     void SetCell(Cell& cell, Cell::State state);
 
-    const Cell& GetCell(int x, int y) const
+    const Cell& GetCell(uint16_t x, uint16_t y) const
     {
         if (x * y > _size)
         {
@@ -111,7 +111,7 @@ public:
         return _board[x + (y * _width)];
     }
 
-    Cell& GetCell(int x, int y)
+    Cell& GetCell(uint16_t x, uint16_t y)
     {
         if (x * y > _size)
         {
@@ -120,9 +120,9 @@ public:
         return _board[x + (y * _width)];
     }
 
-    int CountLiveAndDyingNeighbors(int x, int y);
+    uint16_t CountLiveAndDyingNeighbors(uint16_t x, uint16_t y);
 
-    int CountLiveNotDyingNeighbors(int x, int y);
+    uint16_t CountLiveNotDyingNeighbors(uint16_t x, uint16_t y);
 
     void ApplyNextStateToBoard();
 
@@ -132,9 +132,9 @@ public:
     // but using auto is magic
     void UpdateBoardWithNextState(auto F)
     {
-        for (int y = 0; y < Height(); y++)
+        for (uint16_t y = 0; y < Height(); y++)
         {
-            for (int x = 0; x < Width(); x++)
+            for (uint16_t x = 0; x < Width(); x++)
             {
                 Cell& cc = GetCell(x, y);
                 CountLiveAndDyingNeighbors(x, y);
